Kafka-Producer/requirements.txt
confluent-kafka==3.2